Это объединение 2 коммитов. (#23)
Это 1-е сообщение коммита:

SU4NA-26: Логирование http-запросов

Это сообщение коммита номер #2:

SU4NA-50: Обновление тайтлов пачкой не работает
